get rid of wasmtime dependency entirely
