+++ Cargo.toml
@@ -13,6 +13,10 @@ wasmtime-wasi = "^0.35"
 wasi-common = "^0.35"
 chrono = "0.4.19"
 anyhow = "1.0"
+cap-std= "0.24.1"
+oci-spec = "0.5"
+thiserror = "1.0"
+serde_json = "1.0"
 
 [dev-dependencies]
 tempfile = "3.0"
